Add files via upload
This is the very first project I created an image/advertisement for our showcase day in the Graph courses. I made it after just six weeks of learning Generative AI and Work for Research . It may look simple, but I put in my best effort to create it, so a little applause is welcome!